Fixed bug which was printing extra nodes in path.
I was adding a node to the visited list whenever I was pushing it into a queue. That was the wrong approach because I might find the goal node even before the queue is exhausted. So I had to change my code such that the node is appended to the visited list only when we check whether or not it's our goal node.